requirements.txt
asgiref==3.5.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.0.4
django-heroku==0.3.1
django-redis==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pyparsing==3.0.8
pytz==2022.1
redis==4.3.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
wrapt==1.14.1
